Updated MSI-X handling in Qemu
According to the spec, the P bit should be set whenever an interrupt is requested and that particular vector is masked/being serviced. So whenever that interrupt vector is unmasked, hardware should read the PBA, and generate appropriate interrupts if it's not already serviced/acknowledged. Host acknowledges the same by writing to the CQyHDBL, and as a result hardware should check whether all the Queues associated with that interrupt vector are empty and only then clear the P bit. MSI-x support in Qemu has been updated to reflect the same.