QEMU: a few nvme enhancements
Fixed the set feature number of queues to not actually save this, but return
the maximum number of queues available as will happen on SaffordPeak.

Fixed the namespace utilization when the namespace was formatted with a
different lba size.

Changed BYTES_PER_BLOCK from hard-coded value to a calculation from the lba
size to make it easier to test different sizes; this helped find and fix a Linux
driver block queue issue.