QEMU NVMe: Fultondale stripe boundary
Added nvme device option "fultondale=<1,2,3,4>" to emulate fultondale features,
supporting the vendor unique striping. The device differentiates from standard
QEMU NVME device by overriding the pci device id with a unique one. The
different values for the "fultondale" option are an index for the selected
stripe sizes defined in the spec. IO's that cross the stripe boundary will
result in an error message printed on the terminal.

Signed-off-by: Keith Busch <keith.busch@intel.com>